chore: Cleanup the merge job to outline the logic and make it more clear what's a happening
The idea is to select a primary Playlist, then any failover playlists. The primary playlist channels will be looped through, and we will search for any matching channels in the selected failover Playlists. If matches are found, they will be added to the primary channel as a failover source. The order will be determined by the failover playlist selection, or by resolution when selected.